requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
Jinja2==3.1.2
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
